issue #785 refactor
